refactor(rag): extract answer/search stage helpers for pipeline
